login
login

+++ Ggorrrr/src/com/ggorrrr/web/controller/member/MyBookmarkListController.java
@@ -22,9 +22,8 @@ public class MyBookmarkListController extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		
-		
-		request.setAttribute("list", bookmarkService.getList(Long.parseLong(request.getParameter("id"))));
+
+		request.setAttribute("list", bookmarkService.getList(Integer.parseInt(request.getParameter("id"))));
 		request.getRequestDispatcher("/WEB-INF/view/member/myBookmarkList.jsp").forward(request, response);
 	}
